REFACT: Clean up the code, make it more modular, fix minor issues
